Simplify: 국가 선택 모달 컴포넌트 이동
다른 파일로 분리한 SelectContryModal 컴포넌트는 Start.jsx에만 사용할 것으로 판단하여 코드를 Start.jsx 안에 이동했습니다
