+++ subakFront/src/pages/login/Start.jsx
@@ -1,12 +1,10 @@
 // 시작 화면
 import { react, useState } from 'react';
-import {Image, Text, View, TouchableOpacity} from 'react-native';
+import {Image, Text, View, TouchableOpacity, Modal, TouchableWithoutFeedback} from 'react-native';
+import WheelPicker from 'react-native-wheely';
 
 import shared from '../../styles/shared';
 import styles from '../../styles/start';
-import SelectContryModal from '../../components/SelectCountryModal';
-import FindLocate from './FindLocate';
-import Login from './Login';
 
 const Index = () => {
   // 국가 선택 버튼
@@ -49,6 +47,47 @@ const Index = () => {
               onPress={() => console.log("onPress")}> 로그인</Text>
           </Text>
       </View>
+    </View>
+  );
+};
+
+/**
+ * 국가 선택 모달창
+ * @param {countryIndex}
+ * @returns 
+ */
+const SelectContryModal = props => {
+  var [index, setIndex] = useState(props.countryIndex);
+
+  return (
+    <View style={styles.modalContainer}>
+      <Modal
+        visible={props.openModal}
+        animationType={'slide'}
+        transparent={true}
+        onRequestClose={() => props.setOpenModal(false)}>
+        <TouchableWithoutFeedback onPress={() => props.setOpenModal(false)}>
+          <View style={styles.backDrop} />
+        </TouchableWithoutFeedback>
+
+        <View style={styles.modal}>
+          <Text style={styles.modalText}>국가를 선택해주세요</Text>
+          <WheelPicker
+            selectedIndex={props.countryIndex}
+            options={props.country}
+            onChange={i => setIndex(i)}
+            itemStyle={styles.select}
+            itemTextStyle={styles.itemText}
+          />
+          <TouchableOpacity
+            onPress={() => {
+              props.setCountryIndex(index);
+              props.setOpenModal(false);
+            }}>
+            <Text style={[shared.button, styles.text]}>확인</Text>
+          </TouchableOpacity>
+        </View>
+      </Modal>
     </View>
   );
 };
